media-source/SourceBuffer-appendBuffer-updating.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>SouceBuffer#updating attribute after calling SourceBuffer#appendBuffer(data)</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id="log"></div>
<script>
var video = document.createElement('video');
var contents = {'/media/white.webm' : 'video/webm; codecs="vorbis,vp8"',
                '/media/white.mp4'  : 'video/mp4'};

//check the browser supports the MIME used in this test
function isTypeSupported(mime) {
    if(!MediaSource.isTypeSupported(mime)) {
        this.step(function() {
            assert_unreached("Browser doesn't support the MIME used in this test: " + mime);
        });
        this.done();
        return false;
    }
    return true;
}
function GET(url, processBody) {
    var xhr = new XMLHttpRequest();
    xhr.open('GET', url, true);
    xhr.responseType = 'arraybuffer';
    xhr.send();
    xhr.onload = function(e) {
        if (xhr.status != 200) {
            alert("Unexpected status code " + xhr.status + " for " + url);
            return false;
        }
        processBody(new Uint8Array(xhr.response));
    };
}
function testMedia(file, mime) {
    async_test(function(t) {
        if(!isTypeSupported.bind(t)(mime)) {
            return;
        }
        GET(file, function(data) {
            var mediaSource = new MediaSource();
            mediaSource.addEventListener('sourceopen', function(e) {
                var sourceBuffer = mediaSource.addSourceBuffer(mime);
                var updatings = [];
                sourceBuffer.addEventListener('updatestart', function(e) {
                    updatings.push(sourceBuffer.updating);
                });
                sourceBuffer.addEventListener('update', function(e) {
                    updatings.push(sourceBuffer.updating);
                });
                sourceBuffer.addEventListener('updateend', function(e) {
                    updatings.push(sourceBuffer.updating);
                    t.step(function() {
                        assert_array_equals(updatings,
                                            [false, true, true, false, false],
                                            'Check the transition of the updating attribute');
                    });
                    t.done();
                });
                updatings.push(sourceBuffer.updating);
                sourceBuffer.appendBuffer(data);
                updatings.push(sourceBuffer.updating);
            });
            video.src = window.URL.createObjectURL(mediaSource);
        });
    }, 'SourceBuffer#appendBuffer(data) (' + mime + '): ' +
       'Check the transition of updating attribute.');
}
for(var file in contents) {
    testMedia(file, contents[file]);
}
</script>
</body>
</html>
